Hook up callbacks to IOKit APIs
This seems to have fixed the issue with secure keyboard getting around
the API however it does not seem to be seizing the events as they still
are propagated throughout the system. I belive I am passing in the
correct flag to the API however I am not sure what is going on. Will
look into it more later.